// src/interfaces/Requirement.ts
export interface Requirement {
    courses: string[];
    tech: number;
    creative: number; // "Creative Arts and Humanities"
    history: number; // "History and Cultural Change"
    social: number; // "Social and Behavioral Sciences"
    math: number; // "Mathematics, Natural Sciences, and Technology"
}